CEF Python 53.1 Beta release for Linux and Python 2.7 / 3.4 / 3.5.
Update to Chrome 53.0.2785.101 / CEF 3.2785.1481.gc3a7e2c.

Add drag-image representation of d&d in OSR Kivy (#251).

Create a patch for CEF to expose drag-image representation in
CefDragData.

Make a clean shutdown of CEF in Kivy example. Add a call to CloseBrowser()
in App.on_stop().

Add DragData.GetImage() and HasImage() methods, see API docs.

Add Image object and its methods to get image either as bitmap or png,
see API docs.

Fix issue during Shutdown in cefpython internals
("dictionary changed size during iteration" error).

Fix issue with invalid path to "locales_dir_path" on Python 3.

Fix automate.py. Fix --no-cef-update flag in automate.py. Fix issues
with --force-chromium-update. Fix fetching CEF commit from src/version/.
Print error message when automate.py is run on Python != 2.7.

Remove setup.cfg from Linux installer, this is a remaining from Windows
installer. On Linux no wheels are created.

Include the hello_world.py example in setup installer.

+++ src/linux/compile.py
@@ -56,6 +56,12 @@ if len(sys.argv) > 1 and "--fast" in sys.argv:
 else:
     FAST = False
 
+if len(sys.argv) > 1 and "--kivy" in sys.argv:
+    KIVY = True
+    print("KIVY mode On")
+else:
+    KIVY = False
+
 if len(sys.argv) > 1 and re.search(r"^\d+\.\d+$", sys.argv[1]):
     VERSION = sys.argv[1]
 else:
@@ -187,9 +193,12 @@ if DEBUG:
     os.chdir("./binaries_%s" % BITS)
     subprocess.call("cygdb . --args python-dbg wxpython.py", shell=True)
 else:
-    os.system("rm -rf ./installer/cefpython3-%s-*" % (VERSION,))
-    subprocess.call("cd ./installer/ && python make-setup.py --version %s"
-                    " && cd cefpython3-%s-* && python setup.py install"
-                    " && cd ../../../../examples/"
-                    " && python hello_world.py && cd ../src/linux/"\
-                    % (VERSION,VERSION), shell=True)
+    if KIVY:
+        os.system("python binaries_64bit/kivy_.py")
+    else:
+        os.system("rm -rf ./installer/cefpython3-%s-*" % (VERSION,))
+        subprocess.call("cd ./installer/ && python make-setup.py --version %s"
+                        " && cd cefpython3-%s-* && python setup.py install"
+                        " && cd ../../../../examples/"
+                        " && python hello_world.py && cd ../src/linux/"\
+                        % (VERSION,VERSION), shell=True)
